fixup! User can sign in and log out
