+++ spec/spec_helper.rb
@@ -1,14 +1,12 @@
-
+require "support/features/authentication_helpers"
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
-
+  config.include Features::AuthenticationHelpers, type: :feature
   config.mock_with :rspec do |mocks|
     mocks.verify_partial_doubles = true
   end
-
   config.shared_context_metadata_behavior = :apply_to_host_groups
-
   config.order = :random
 end
